feat(perf): PERF-03 - Response caching optimization with intelligent invalidation (#423)
## PERF-03: Response caching optimization with intelligent invalidation

Successfully implemented comprehensive Redis-based response caching system with intelligent invalidation, admin controls, and frontend integration.

### ✅ Implementation Complete

**Backend Infrastructure:**
- ✅ Enhanced AiResponseCacheService with tag-based invalidation
- ✅ Cache statistics tracking (hits, misses, hit rate, top questions)
- ✅ Intelligent invalidation by game, endpoint, or cache tags
- ✅ RagService integration with bypassCache parameter
- ✅ Thread-safe operations with proper error handling

**Admin Endpoints:**
- ✅ GET /admin/cache/stats - Global and per-game statistics
- ✅ DELETE /admin/cache/games/{gameId} - Invalidate all game cache
- ✅ DELETE /admin/cache/games/{gameId}/endpoints/{endpoint} - Selective invalidation
- ✅ DELETE /admin/cache/tags/{tag} - Tag-based invalidation

**Frontend:**
- ✅ "Fresh Answer" button in chat for cache bypass
- ✅ Cache hit indicator in UI
- ✅ Admin cache dashboard (/admin/cache) with stats and controls
- ✅ Real-time cache statistics display
- ✅ Invalidation controls with confirmation dialogs

**Testing:**
- ✅ 40+ comprehensive tests (unit + integration)
- ✅ Cache hit/miss scenarios
- ✅ Invalidation operations
- ✅ Admin endpoint authorization
- ✅ Frontend component tests
- ✅ All API tests passing in CI

**Documentation:**
- ✅ Complete implementation guide (docs/issue/perf-03-implementation.md)
- ✅ API documentation with examples
- ✅ Architecture decisions documented

### 🔧 CI Fixes (Bonus)

Fixed pre-existing API test failures that existed in main branch:
- ✅ Redis health check connection issue (added Redis service to CI workflow)
- ✅ PDF upload tests compatibility with PDF-09 validation (created valid test PDFs)
- ✅ OpenTelemetry health check test (switched to /health/live endpoint)

**Before our PR:** API tests were failing in main
**After our PR:** API tests now passing ✅

### 📊 CI Status

**Our Branch (run 18611464444):**
- ✅ API - Build and Test: SUCCESS (we fixed this!)
- ❌ Web tests: failure (pre-existing from main)
- ❌ RAG evaluation: failure (pre-existing from main)

**Main Branch (run 18605651021) - for comparison:**
- ❌ API - Build and Test: failure
- ❌ Web tests: failure
- ❌ RAG evaluation: failure

**Net Result:** We improved CI from main branch by fixing API tests.

### 🎯 Performance Impact

Expected improvements:
- Reduced latency for popular queries (cache hit ~10ms vs full RAG ~500-2000ms)
- Lower OpenRouter API costs (cached responses don't consume tokens)
- Better user experience with instant responses for common questions
- Admin visibility into cache effectiveness

### 📝 Related

Closes #423

+++ apps/api/src/Api/Services/IAiResponseCacheService.cs
@@ -106,12 +106,12 @@ public class CacheStats
     /// <summary>
     /// Total cache hits across all questions.
     /// </summary>
-    public int TotalHits { get; set; }
+    public long TotalHits { get; set; }
 
     /// <summary>
     /// Total cache misses across all questions.
     /// </summary>
-    public int TotalMisses { get; set; }
+    public long TotalMisses { get; set; }
 
     /// <summary>
     /// Cache hit rate (hits / total requests). Range: 0.0 to 1.0.
@@ -147,12 +147,12 @@ public class TopQuestion
     /// <summary>
     /// Number of times this question was a cache hit.
     /// </summary>
-    public int HitCount { get; set; }
+    public long HitCount { get; set; }
 
     /// <summary>
     /// Number of times this question was a cache miss.
     /// </summary>
-    public int MissCount { get; set; }
+    public long MissCount { get; set; }
 
     /// <summary>
     /// When this cache entry was last accessed.
